feat(mercadoPago): implementacion de pasarela de pago y metodo para crear solo con el id de la cita

+++ src/main/java/com/veterinaria/veterinaria_comoreyes/entity/VeterinaryService.java
@@ -28,7 +28,7 @@ public class    VeterinaryService extends EntityWithStatus{
 
     private String description;
 
-    private float price;
+    private Double price;
 
     private Integer duration;
